Add health checks for schema migration

+++ misk-hibernate/src/test/kotlin/misk/hibernate/SchemaMigratorTest.kt
@@ -1,42 +1,49 @@
 package misk.hibernate
 
-import com.google.common.util.concurrent.AbstractIdleService
+import com.google.common.collect.Iterables.getOnlyElement
+import com.google.inject.util.Modules
 import misk.MiskTestingServiceModule
-import misk.ServiceModule
 import misk.config.Config
 import misk.config.MiskConfig
 import misk.environment.Environment
-import misk.inject.KAbstractModule
-import misk.inject.asSingleton
-import misk.inject.keyOf
+import misk.environment.EnvironmentModule
 import misk.jdbc.DataSourceConfig
 import misk.jdbc.DataSourceService
-import misk.jdbc.PingDatabaseService
 import misk.resources.ResourceLoader
 import misk.testing.MiskTest
 import misk.testing.MiskTestModule
-import misk.vitess.StartVitessService
 import org.assertj.core.api.Assertions.assertThat
-import org.hibernate.SessionFactory
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import javax.inject.Inject
 import javax.inject.Provider
-import javax.inject.Singleton
 import javax.persistence.PersistenceException
-import javax.sql.DataSource
 import kotlin.test.assertFailsWith
 
-@MiskTest(startService = true)
+@MiskTest(startService = false)
 internal class SchemaMigratorTest {
   val defaultEnv = Environment.TESTING
   val config = MiskConfig.load<RootConfig>("test_schemamigrator_app", defaultEnv)
 
-  @Singleton
-  class DropTablesService @Inject constructor() : AbstractIdleService() {
-    @Inject @Movies lateinit var sessionFactoryProvider: Provider<SessionFactory>
+  @MiskTestModule
+  val module = Modules.combine(
+      EnvironmentModule(Environment.TESTING),
+      MiskTestingServiceModule(),
+      HibernateModule(Movies::class, config.data_source)
+  )
 
-    override fun startUp() {
-      sessionFactoryProvider.get().openSession().use { session ->
+  @Inject lateinit var resourceLoader: ResourceLoader
+  @Inject @Movies lateinit var transacter: Provider<Transacter>
+  @Inject @Movies lateinit var dataSourceService: DataSourceService
+  @Inject @Movies lateinit var sessionFactoryService: SessionFactoryService
+  @Inject @Movies lateinit var schemaMigrator: SchemaMigrator
+  @Inject @Movies lateinit var schemaMigratorService: SchemaMigratorService
+
+  @AfterEach
+  internal fun tearDown() {
+    if (sessionFactoryService.isRunning) {
+      sessionFactoryService.get().openSession().use { session ->
         session.doReturningWork { connection ->
           val statement = connection.createStatement()
           statement.addBatch("DROP TABLE IF EXISTS schema_version")
@@ -50,69 +57,16 @@ internal class SchemaMigratorTest {
         }
       }
     }
-
-    override fun shutDown() {}
   }
 
-  @MiskTestModule
-  val module = object : KAbstractModule() {
-    override fun configure() {
-      install(MiskTestingServiceModule())
-
-      bind(keyOf<StartVitessService>(Movies::class)).toInstance(
-          StartVitessService(Movies::class, Environment.TESTING, config.data_source))
-      install(ServiceModule<StartVitessService>(Movies::class))
-
-      bind(keyOf<PingDatabaseService>(Movies::class)).toInstance(
-          PingDatabaseService(config.data_source, defaultEnv))
-      install(ServiceModule<PingDatabaseService>(Movies::class)
-          .dependsOn<StartVitessService>(Movies::class))
-
-      install(ServiceModule<DropTablesService>())
-
-      val dataSourceService =
-          DataSourceService(Movies::class, config.data_source, defaultEnv, emptySet())
-      bind(keyOf<DataSourceService>(Movies::class)).toInstance(dataSourceService)
-      bind<DataSource>().annotatedWith<Movies>().toProvider(dataSourceService)
-      install(ServiceModule<DataSourceService>(Movies::class)
-          .dependsOn<PingDatabaseService>(Movies::class))
-
-      val injectorServiceProvider = getProvider(HibernateInjectorAccess::class.java)
-      val sessionFactoryServiceKey = keyOf<SessionFactoryService>(Movies::class)
-      bind(sessionFactoryServiceKey).toProvider(Provider<SessionFactoryService> {
-        SessionFactoryService(Movies::class, config.data_source, dataSourceService,
-            injectorServiceProvider.get())
-      }).asSingleton()
-      bind<SessionFactory>().annotatedWith<Movies>().toProvider(sessionFactoryServiceKey)
-      val sessionFactoryKey = keyOf<SessionFactory>(Movies::class)
-      val sessionFactoryProvider = getProvider(sessionFactoryKey)
-
-      bind(keyOf<TransacterService>(Movies::class)).to(keyOf<SessionFactoryService>(Movies::class))
-      install(ServiceModule<TransacterService>(Movies::class)
-          .enhancedBy<DropTablesService>()
-          .dependsOn<DataSourceService>(Movies::class))
-
-      val transacterKey = keyOf<Transacter>(Movies::class)
-      bind(transacterKey).toProvider(object : Provider<Transacter> {
-        @Inject lateinit var queryTracingListener: QueryTracingListener
-        override fun get(): RealTransacter = RealTransacter(
-            Movies::class,
-            sessionFactoryProvider,
-            config.data_source,
-            queryTracingListener,
-            null
-        )
-      }).asSingleton()
-    }
+  @BeforeEach internal fun setUp() {
+    dataSourceService.startAsync()
+    dataSourceService.awaitRunning()
+    sessionFactoryService.startAsync()
+    sessionFactoryService.awaitRunning()
   }
-
-  @Inject lateinit var resourceLoader: ResourceLoader
-  @Inject @Movies lateinit var transacter: Provider<Transacter>
 
   @Test fun initializeAndMigrate() {
-    val schemaMigrator =
-        SchemaMigrator(Movies::class, resourceLoader, transacter, config.data_source)
-
     val mainSource = config.data_source.migrations_resources!![0]
     val librarySource = config.data_source.migrations_resources!![1]
 
@@ -197,8 +151,6 @@ internal class SchemaMigratorTest {
   }
 
   @Test fun requireAllWithMissingMigrations() {
-    val schemaMigrator =
-        SchemaMigrator(Movies::class, resourceLoader, transacter, config.data_source)
     schemaMigrator.initialize()
 
     resourceLoader.put("${config.data_source.migrations_resources!![0]}/v1001__foo.sql", """
@@ -217,9 +169,6 @@ internal class SchemaMigratorTest {
   }
 
   @Test fun errorOnDuplicateMigrations() {
-    val schemaMigrator =
-        SchemaMigrator(Movies::class, resourceLoader, transacter, config.data_source)
-
     resourceLoader.put("${config.data_source.migrations_resources!![0]}/v1001__foo.sql", """
         |CREATE TABLE table_1 (name varchar(255))
         |""".trimMargin())
@@ -248,6 +197,22 @@ internal class SchemaMigratorTest {
     assertThat(namespacedMigrationOrNull("foo/migrations/v100__.sql")).isNull()
     assertThat(namespacedMigrationOrNull("foo/migrations/v100__.sql")).isNull()
     assertThat(namespacedMigrationOrNull("foo/luv1__franklin.sql")).isNull()
+  }
+
+  @Test fun healthChecks() {
+    resourceLoader.put("${config.data_source.migrations_resources!![0]}/v1002__movies.sql", """
+        |CREATE TABLE table_2 (name varchar(255))
+        |""".trimMargin())
+    resourceLoader.put("${config.data_source.migrations_resources!![0]}/v1001__movies.sql", """
+        |CREATE TABLE table_1 (name varchar(255))
+        |""".trimMargin())
+
+    schemaMigratorService.startAsync()
+    schemaMigratorService.awaitRunning()
+
+    assertThat(getOnlyElement(schemaMigratorService.status().messages)).isEqualTo(
+        "SchemaMigratorService: Movies is migrated: " +
+            "MigrationState(shards={keyspace/0=(all 2 migrations applied)})")
   }
 
   private fun tableExists(table: String): Boolean {
